+++ Fabric/src/main/java/com/christofmeg/justenoughbreeding/emi/BreedingCategoryEMI.java
@@ -23,6 +23,7 @@ import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
+import net.minecraft.world.item.crafting.Ingredient;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
@@ -69,7 +70,8 @@ public class BreedingCategoryEMI implements EmiRecipe {
     @Override
     public List<EmiIngredient> getCatalysts() {
         return List.of(EmiIngredient.of(recipe.breedingCatalyst),
-                EmiIngredient.of(recipe.extraInputStack)
+                EmiIngredient.of(recipe.extraInputStack),
+                EmiIngredient.of(Ingredient.of(recipe.spawnEgg))
                 );
     }
 
@@ -80,6 +82,7 @@ public class BreedingCategoryEMI implements EmiRecipe {
             for (ItemStack item : recipe.resultItemStack.getItems()) {
                 list.add(EmiStack.of(item));
             }
+            list.add(EmiStack.of(recipe.spawnEgg));
         }
         return list;
     }