+++ NeoForge/changelog.md
@@ -1,2 +1,5 @@
-1.2.2:
-- Remove optional dependency to REIPC, it was dropped on 1.20.1
+1.2.4:
+- Made breeding category searchable with spawn eggs
+
+1.2.3:
+- Added russian translation by [ImpOfficer](https://github.com/ImpOfficer)